Simplify the event array model
Include VP events too. This is somewhat redundant and
repetitive information, but leaving out VP would lead
to more complex and harder-to-read code.
It will also be interesting to see whether there actually are
occurrences where VP was not triggered along with other events
for some second+vehicle.
Also simplify the index: the WHERE clause can disable index scan
even if it were needed, and we should only optimize the index
if it is proven to be too slow with VPs included.