Use custom set_modified_at() trigger function
Use the custom function instead of the
moddatetime extension that is not supported
by Azure SQL for Postgres.

NOTE: While the moddatetime function took the
target column name as a parameter, this custom one
is hard-coded to support `modified_at` column only!
This in turn ensures that this column is named
the same way wherever used in this database.